// QuickApp/ClientApp/src/app/models/svc.model.ts
// =============================
// Email: info@ebenmonney.com
// www.ebenmonney.com/templates
// =============================


export class Svc {

  id: number;
  domainId: number;
  accountId: number;
  country: string;
  serviceDeliveryManager: string;
  accountManager: string;
  date: Date;
  readableDate: string;
  newDate: string;
  quoteFTL: string;
  po: string;
  client: string;
  field: string;
  well: string;
  au: string;
  ac: string;
  portfolio: string;
  subPortfolio: string;
  masterCode: string;
  currency: string;
  fxRate: number;
  comment: string;
  technicalLead: string;
  changePointTask: string;
  rofo: number;
  imf: number;
  mmf: number;
  sentToInvoice: number;
  revenue: number;
  invocieNumber: string;
  cost: number;
  costReceived: number;
  costType: string;
  glAccount: string;
  costDescription: string;
  LockedBy: string;
}